+++ app.json
@@ -1,0 +1,12 @@
+{
+    "addons": [
+        "heroku-postgresql:hobby-dev"
+    ],
+    "environments": {
+        "review": {
+            "addons": [
+                "heroku-postgresql:hobby-dev"
+            ]
+        }
+    }
+}
