--- backend/.ebextensions/nodecommand.config
@@ -1,3 +1,0 @@
-option_settings:
-  aws:elasticbeanstalk:container:nodejs:
-    NodeCommand: "npm start"
