+++ backend/.env.development
@@ -1,2 +1,2 @@
 PORT=8760
-DEV_DATABASE_URL=postgres://ruth:ruth@localhost:15432/ruth
+DATABASE_URL=postgres://ruth:ruth@localhost:15432/ruth
