+++ backend/.gitignore
@@ -60,7 +60,5 @@ typings/
 # next.js build output
 .next
 
-# Elastic Beanstalk Files
-.elasticbeanstalk/*
-!.elasticbeanstalk/*.cfg.yml
-!.elasticbeanstalk/*.global.yml
+# build
+build/
