+++ backend/src/database/config/config.js
@@ -2,11 +2,7 @@ import '~/bootstrap/env';
 
 module.exports = {
   development: {
-    url: process.env.DEV_DATABASE_URL,
-    dialect: 'postgres',
-  },
-  test: {
-    url: process.env.TEST_DATABASE_URL,
+    url: process.env.DATABASE_URL,
     dialect: 'postgres',
   },
   production: {
