+++ backend/src/server.js
@@ -1,6 +1,7 @@
 import express, { json, urlencoded } from 'express';
 import cookieParser from 'cookie-parser';
 import morgan from 'morgan';
+import * as path from 'path';
 
 import indexRouter from './routes';
 import logger from '~/logger';
@@ -12,8 +13,14 @@ app.use(json());
 app.use(urlencoded({ extended: false }));
 app.use(cookieParser());
 
-app.use('/', indexRouter);
+app.use('/api', indexRouter);
+if (process.env.NODE_ENV === 'production') {
+  console.log("dir", __dirname);
+  app.use(express.static(path.join(__dirname, './frontend')));
+}
+
 app.use(logger.errorHandler());
+
 app.get('*', (req, res) => {
   logger.info(`Request not found ${req.url}`);
   res.status(404).send('Not found');
