+++ package.json
@@ -4,6 +4,7 @@
   "description": "Presentacionse roots",
   "main": "index.js",
   "scripts": {
+    "start": "cd backend/build && node index.js",
     "install:frontend": "cd frontend && npm install",
     "install:backend": "cd backend && npm install",
     "postinstall": "concurrently \"npm run install:backend\" \"npm run install:frontend\"",
@@ -13,6 +14,10 @@
     "test:frontend": "cd frontend && npm run test",
     "test:backend": "cd backend && npm run test",
     "test": "npm run test:frontend && npm run test:backend",
+    "build:frontend": "cd frontend && npm run build",
+    "build:backend": "cd backend && npm run build",
+    "build:pegadito": "cp -r frontend/build backend/build/frontend/",
+    "build": "npm run build:frontend && npm run build:backend && npm run build:pegadito",
     "lint:frontend": "cd frontend && npm run lint",
     "lint:backend": "cd backend && npm run lint",
     "lint": "npm run lint:frontend && npm run lint:backend"
@@ -26,6 +31,10 @@
     "presentacion",
     "debate"
   ],
+  "engines": {
+      "node": "10.16.0",
+      "npm": "6.9.0"
+  },
   "author": "",
   "license": "MIT",
   "devDependencies": {
